requirements.txt
alembic==1.10.3
anyio==3.6.2
awscli==1.27.115
boto3==1.26.114
botocore==1.29.115
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.6
dgl==1.0.1
docker==6.0.1
docutils==0.16
entrypoints==0.4
fastapi==0.95.0
filelock==3.11.0
Flask==2.2.3
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
minio==7.1.14
mlflow==2.2.2
mpmath==1.3.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
protobuf==4.22.3
psutil==5.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.28.2
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.9
sqlparse==0.4.3
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torch_geometric==2.3.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
zipp==3.15.0
timeout
